requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
authlib==0.15.4
    # via drf-oidc-auth
cachetools==4.2.4
    # via django-helusers
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.6
    # via requests
cryptography==35.0.0
    # via
    #   authlib
    #   drf-oidc-auth
    #   social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
deprecation==2.1.0
    # via django-helusers
django==3.2.8
    # via
    #   -r requirements.in
    #   django-helusers
    #   djangorestframework
    #   drf-oidc-auth
django-environ==0.7.0
    # via -r requirements.in
django-helusers==0.7.0
    # via -r requirements.in
djangorestframework==3.12.4
    # via
    #   -r requirements.in
    #   drf-oidc-auth
drf-oidc-auth==2.0.0
    # via -r requirements.in
ecdsa==0.17.0
    # via python-jose
idna==3.2
    # via requests
oauthlib==3.1.1
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.0
    # via deprecation
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.20
    # via cffi
pyjwt==2.2.0
    # via social-auth-core
pyparsing==2.4.7
    # via packaging
python-jose==3.3.0
    # via django-helusers
python3-openid==3.2.0
    # via social-auth-core
pytz==2021.3
    # via django
requests==2.26.0
    # via
    #   django-helusers
    #   drf-oidc-auth
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.0
    # via social-auth-core
rsa==4.7.2
    # via python-jose
six==1.16.0
    # via ecdsa
social-auth-app-django==5.0.0
    # via -r requirements.in
social-auth-core==4.1.0
    # via social-auth-app-django
sqlparse==0.4.2
    # via django
typing-extensions==3.10.0.2
    # via asgiref
urllib3==1.26.7
    # via requests
